requirements_global.txt
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
pyzmq==27.1.0
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tomlkit==0.13.3
tornado==6.5.4
traitlets==5.14.3
trove-classifiers==2025.11.14.15
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.14
zstandard==0.25.0
